+++ stingray/simulator/transfer.py
@@ -4,7 +4,7 @@ import numpy as np
 from stingray.io import read, write
 
 """
-Implementation of time and energy averaged responses from 2-d 
+Implementation of time and energy averaged responses from 2-d
 transfer functions.
 """
 
@@ -68,8 +68,8 @@ class TransferFunction(object):
 
     def time_response(self, e0=None, e1=None):
         """
-        Form an energy-averaged/time-resolved response of 2-d transfer 
-        function. 
+        Form an energy-averaged/time-resolved response of 2-d transfer
+        function.
 
         Returns
         -------
@@ -82,13 +82,13 @@ class TransferFunction(object):
         e1: int
             end value of energy interval to be averaged
         """
-        
+
         # Set start and stop values
         if e0 is None:
             start = 0
         else:
             start = int(self.estart + e0/self.de)
-        
+
         if e1 is None:
             stop = len(self.data[:][0]) - 1
         else:
@@ -108,14 +108,14 @@ class TransferFunction(object):
 
     def energy_response(self):
         """
-        Form a time-averaged/energy-resolved response of 2-d transfer function. 
+        Form a time-averaged/energy-resolved response of 2-d transfer function.
 
         Returns
         -------
         time : numpy.ndarray
             time-averaged/energy-resolved response of 2-d transfer function
         """
-        
+
         self.energy = np.mean(self.data, axis=1)
 
     def plot(self, response='2d', save=False, filename=None, show=False):
@@ -145,7 +145,7 @@ class TransferFunction(object):
         fig = plt.figure()
 
         if response == 'time':
-            t = np.linspace(self.tstart, len(self.data[0])*self.dt, 
+            t = np.linspace(self.tstart, len(self.data[0])*self.dt,
                             len(self.data[0]))
             figure = plt.plot(t, self.time)
             plt.xlabel('Time')
@@ -153,13 +153,13 @@ class TransferFunction(object):
             plt.title('Time-resolved Response')
 
         elif response == 'energy':
-            e = np.linspace(self.estart, len(self.data[:])*self.de, 
+            e = np.linspace(self.estart, len(self.data[:])*self.de,
                             len(self.data[:]))
             figure = plt.plot(e, self.energy)
             plt.xlabel('Energy')
             plt.ylabel('Flux')
             plt.title('Energy-resolved Response')
-            
+
         elif response == '2d':
             figure = plt.imshow(self.data, interpolation='nearest',
                                 cmap='Oranges', origin='lower')
@@ -177,12 +177,12 @@ class TransferFunction(object):
                 plt.savefig('out.png')
             else:
                 plt.savefig(filename)
-        
+
         if show:
             plt.show()
         else:
             plt.close()
-        
+
     @staticmethod
     def read(filename, format_='pickle'):
         """
@@ -198,7 +198,7 @@ class TransferFunction(object):
         data : class instance
             `TransferFunction` object
         """
-        
+
         object = read(filename, format_)
 
         if format_ == 'pickle':
@@ -209,7 +209,7 @@ class TransferFunction(object):
 
     def write(self, filename, format_='pickle'):
         """
-        Writes a transfer function to 'pickle' file. 
+        Writes a transfer function to 'pickle' file.
 
         Parameters
         ----------
@@ -224,13 +224,13 @@ class TransferFunction(object):
             raise KeyError("Format not understood.")
 
 """
-Implementation of artificial methods to create energy-averaged 
+Implementation of artificial methods to create energy-averaged
 responses for quick testing.
 """
 
 def simple_ir(dt=0.125, start=0, width=1000, intensity=1):
     """
-    Construct a simple impulse response using start time, 
+    Construct a simple impulse response using start time,
     width and scaling intensity.
     To create a delta impulse response, set width to 1.
 
@@ -241,10 +241,10 @@ def simple_ir(dt=0.125, start=0, width=1000, intensity=1):
 
     start : int
         start time of impulse response
-    
+
     width : int
         width of impulse response
-    
+
     intensity : float
         scaling parameter to set the intensity of delayed emission
         corresponding to direct emission.
@@ -275,22 +275,22 @@ def relativistic_ir(dt=0.125, t1=3, t2=4, t3=10, p1=1, p2=1.4, rise=0.6, decay=0
 
     t1 : int
         primary peak time
-    
+
     t2 : int
         secondary peak time
-    
+
     t3 : int
         end time
-    
+
     p1 : float
         value of primary peak
-    
+
     p2 : float
         value of secondary peak
-    
+
     rise : float
         slope of rising exponential from primary peak to secondary peak
-    
+
     decay : float
         slope of decaying exponential from secondary peak to end time
 
@@ -308,16 +308,16 @@ def relativistic_ir(dt=0.125, t1=3, t2=4, t3=10, p1=1, p2=1.4, rise=0.6, decay=0
     h_primary = np.append(np.zeros(int(t1/dt)), p1)
 
     # Create a rising exponential of user-provided slope
-    x = np.linspace(t1/dt, t2/dt, (t2-t1)/dt)
+    x = np.linspace(t1/dt, t2/dt, int((t2-t1)/dt))
     h_rise = np.exp(rise*x)
-    
+
     # Evaluate a factor for scaling exponential
     factor = np.max(h_rise)/(p2-p1)
     h_secondary = (h_rise/factor) + p1
 
     # Create a decaying exponential until the end time
-    x = np.linspace(t2/dt, t3/dt, (t3-t2)/dt)
-    h_decay = (np.exp((-decay)*(x-4/dt))) 
+    x = np.linspace(t2/dt, t3/dt, int((t3-t2)/dt))
+    h_decay = (np.exp((-decay)*(x-4/dt)))
 
     # Add the three responses
     h = np.append(h_primary, h_secondary)
